requirements.txt
# Django and Core Packages
Django==5.1.3
mysqlclient==2.1.1
python-telegram-bot==13.15

# Django Admin Extensions
django-admin-rangefilter==0.13.2

# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Celery and Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Google API Clients
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-analytics-data==0.18.6

# Utilities
requests==2.31.0
beautifulsoup4==4.13.3
dnspython==2.4.2
python-dotenv==1.0.0

# Existing Dependencies (keep if already in use)
# APScheduler will be managed by python-telegram-bot dependency
pytz==2024.2
